Bug Fix: Added an ID for each player which now accurately reflects who's what player number.
All player handling should be fixed: players are removed from the game when they go bankrupt on the correct turn, and the game notifies the players and restarts when only one person is left.

+++ Monopoly2/src/test/TestPlayer.java
@@ -31,7 +31,7 @@ public class TestPlayer {
 
     @BeforeMethod
     public void setUp() throws IOException {
-        player1 = new Player(1);
+        player1 = new Player((byte) 1);
     }
 
     @AfterMethod
